python/requirements.txt
openai
aredis